Crate key_management

Modules

json

Structs

Key
A Key, this contains a key_info, address, and public_key which holds the key type and private key

KeyInfo
KeyInfo struct, this contains the type of key (stored as a string) and the private key. note how the private key is stored as a byte vector

KeyStore
KeyStore struct, this contains a HashMap that is a set of KeyInfos resolved by their Address

PersistentKeyInfo

Wallet
This is a Wallet, it contains 2 HashMaps:

Enums

EncryptedKeyStoreError

Error

KeyStoreConfig

Constants

ENCRYPTED_KEYSTORE_NAME

KEYSTORE_NAME

Traits

Store
KeyStore struct, this contains a HashMap that is a set of KeyInfos resolved by their Address

Functions

export_key_info
Return keyInfo for given Address in KeyStore

find_key
Return Key corresponding to given Address in KeyStore

generate
Generate a new private key

generate_key
Generate new Key of given SignatureType

get_default
Return the default Address for KeyStore

import
Import KeyInfo into KeyStore

list_addrs
Return Vec of Addresses sorted by their string representation in KeyStore

new_address
Return a new Address that is of a given SignatureType and uses the supplied public_key

sign
Sign takes in SignatureType, private key and message. Returns a Signature for that message

to_public
Return the public key for a given private_key and SignatureType

try_find